nd4j version -> stable

+++ deeplearning4j-scaleout/deeplearning4j-scaleout-akka/src/test/java/org/deeplearning4j/scaleout/runner/TestDistributed.java
@@ -66,9 +66,11 @@ public class TestDistributed {
 
     @Test
     public void testNextOne() {
-        Serializable s = SerializationUtils.readObject(finalFile);
-        assumeNotNull(s);
-        finalFile.delete();
+        if(finalFile.exists()) {
+            Serializable s = SerializationUtils.readObject(finalFile);
+            assumeNotNull(s);
+            finalFile.delete();
+        }
 
     }
